+++ src/libp2p_bitswap/tests/request_manager.rs
@@ -104,8 +104,8 @@ mod tests {
         let peer_id = PeerId::from(id_keys.public());
         let transport = tcp::tokio::Transport::default()
             .upgrade(core::upgrade::Version::V1)
-            .authenticate(noise::NoiseAuthenticated::xx(&id_keys)?)
-            .multiplex(yamux::YamuxConfig::default())
+            .authenticate(noise::Config::new(&id_keys)?)
+            .multiplex(yamux::Config::default())
             .timeout(TIMEOUT)
             .boxed();
         let behaviour = BitswapBehaviour::new(&[b"/test/ipfs/bitswap/1.0.0"], Default::default());